+++ backend/src/main/java/ua/edu/chnu/awards/user/entity/RoleType.java
@@ -1,0 +1,14 @@
+package ua.edu.chnu.awards.user.entity;
+
+/**
+ * Roles a user can hold within an organisation.
+ */
+public enum RoleType {
+    EMPLOYEE,
+    FACULTY_SECRETARY,
+    DEAN,
+    RECTOR_SECRETARY,
+    RECTOR,
+    SYSTEM_ADMIN,
+    GDPR_OFFICER
+}